+++ tools/scaffolding/scaffolder.go
@@ -10,7 +10,6 @@ import (
 	"os/exec"
 	"os/user"
 	"path/filepath"
-	"reflect"
 	"strings"
 	"text/template"
 )
@@ -28,14 +27,14 @@ var sanitizeProto3Identifier = strings.NewReplacer(
 )
 
 type workflowTemplateValues struct {
-	Name             string
-	PackageName      string
-	Description      string
-	DeveloperName    string
-	DeveloperEmail   string
-	URLRoot          string
-	URLPath          string
-	HelloWorldModule string
+	Name           string
+	PackageName    string
+	Description    string
+	DeveloperName  string
+	DeveloperEmail string
+	URLRoot        string
+	URLPath        string
+	WizardTemplate bool
 }
 
 type gatewayTemplateValues struct {
@@ -115,11 +114,11 @@ func getFrontendPluginTemplateValues() (*workflowTemplateValues, string) {
 
 	data := &workflowTemplateValues{}
 
-	data.HelloWorldModule = "hello-wizard"
+	data.WizardTemplate = true
 
 	wizard := promptOrDefault("Is this a wizard workflow?", "Y/n")
 	if !strings.HasPrefix(strings.ToLower(wizard), "y") {
-		data.HelloWorldModule = "hello-world"
+		data.WizardTemplate = false
 	}
 
 	data.Name = strings.Title(promptOrDefault("Enter the name of this workflow", "Hello World"))
@@ -306,8 +305,6 @@ func main() {
 	defer os.RemoveAll(tmpout)
 	log.Println("Using tmpdir", tmpout)
 
-	helloWorldModule := string(reflect.Indirect(reflect.ValueOf(data)).FieldByName("HelloWorldModule").String())
-
 	// Walk files and template them.
 	err = filepath.Walk(templateRoot, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
@@ -322,13 +319,6 @@ func main() {
 				err := os.MkdirAll(filepath.Join(tmpout, relpath), 0755)
 				return err
 			}
-			return nil
-		}
-
-		if relpath == "src/hello-wizard.tsx" && helloWorldModule == "hello-world" {
-			return nil
-		}
-		if relpath == "src/hello-world.tsx" && helloWorldModule == "hello-wizard" {
 			return nil
 		}
