+++ tools/scaffolding/templates/frontend/package.json
@@ -18,7 +18,9 @@
     "test:watch": "yarn run test --watch"
   },
   "dependencies": {
+    {{- if .WizardTemplate}}
     "@clutch-sh/wizard": "^2.0.0-beta",
+    {{- end}}
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
